block mentions in say command

+++ src/main/java/de/eldoria/shepard/commandmodules/standalone/fun/Say.java
@@ -7,7 +7,12 @@ import de.eldoria.shepard.commandmodules.argument.SubCommand;
 import de.eldoria.shepard.commandmodules.command.Executable;
 import de.eldoria.shepard.commandmodules.util.CommandCategory;
 import de.eldoria.shepard.localization.enums.commands.GeneralLocale;
+import de.eldoria.shepard.localization.enums.commands.fun.SayLocale;
+import de.eldoria.shepard.messagehandler.MessageSender;
 import de.eldoria.shepard.wrapper.MessageEventDataWrapper;
+import net.dv8tion.jda.api.entities.Role;
+
+import java.util.List;
 
 import static de.eldoria.shepard.localization.enums.commands.fun.SayLocale.A_SAY;
 import static de.eldoria.shepard.localization.enums.commands.fun.SayLocale.DESCRIPTION;
@@ -30,9 +35,16 @@ public class Say extends Command implements Executable {
 
     @Override
     public void execute(String label, String[] args, MessageEventDataWrapper messageContext) {
-        String message = ArgumentParser.getMessage(args, 0).replace("@", "");
 
-        messageContext.getTextChannel().sendMessage(message).queue();
+        List<Role> mentionedRoles = messageContext.getMessage().getMentionedRoles();
+        boolean everyone = messageContext.getMessage().mentionsEveryone();
+
+        if (!mentionedRoles.isEmpty() || everyone) {
+            MessageSender.sendMessage(SayLocale.A_MENTION.tag, messageContext.getTextChannel());
+            return;
+        }
+
+        messageContext.getTextChannel().sendMessage(ArgumentParser.getMessage(args, 0)).queue();
 
         if (label.equalsIgnoreCase("sayd")) {
             messageContext.getMessage().delete().queue();
